feat: added to TVL endpoint the historical TVL
