+++ services/vaults/holdings/save/handler.js
@@ -21,6 +21,7 @@ const { getHoldings, getPoolTotalSupply, getEarnHoldings } = require('./getHoldi
 const oracle = require('../../../../utils/priceFeed');
 const getVirtualPrice = require('../../apy/save/handler');
 const _ = require('lodash');
+
 const db = dynamodb.doc;
 const Web3 = require('web3');
 
@@ -64,7 +65,6 @@ const readVault = async (vault) => {
         vault.symbol === 'crvBTC' ||
         vault.symbol === '3Crv'
     ){
-      const symbol = vault.symbol;
       const pool = _.find(pools, { symbol });
       const currentBlockNbr = await web3.eth.getBlockNumber();
       const virtualPriceCurrent = await getVirtualPrice.getVirtualPrice(
